awscli v2 support: use 'aws ecr get-login-password'
For awscli >= 1.17.10 we now use `aws ecr get-login-password` instead of
the deprecated (removed in 2.0.0) `aws ecr get-login`.  As a result, we
need to build the registry address, which means determining the AWS
region and account ID.

If the AWS region is not specified in the existing plugin config
options, AWS_DEFAULT_REGION is used, which default to us-east-1.

If the AWS account ID is not specified in the existing plugin config
options, it is (hopefully) found with `aws sts get-caller-identity`.