Refer to ECR URLs using "https://"
This may be a docker bug, or may not be. When you have the ECR
docker-credential helper configured for an account/region combo,
weirldy attempting a `docker login` to that same account/region will
result in an error[1].

However, if we simply prefix the ECR URL with `https`, this error goes
away[2]. I don't believe using this prefix causes any issues otherwise,
and is actually inline with the previous functionality of `v2.0.0` where
it was doing an `eval` of the string returned by `aws ecr get-login`,
which did refer to ECR URLs with that prefix.

Indeed doing an ECR login when you have a credential helper configured
is quite redundant, but we'd rather succeed and then have docker not use
the credentials anyway than fail the build completely due to docker
"failing to store the credentials".

[1]:
```
Error saving credentials: error storing credentials - err: exit status 1, out: `not implemented`
```

[2]:
```
WARNING! Your password will be stored unencrypted in /Users/danial.pearce/.docker/config.json.
Configure a credential helper to remove this warning. See
https://docs.docker.com/engine/reference/commandline/login/#credentials-store

Login Succeeded
```