Fix AWS CLI version parsing
We've had a report of a bug where `workspace` was the returned part of the string at index 1 from the sub-command of `current="$(aws --version 2>&1 | awk -F'[/ ]' '{print $2}')"`

This introduces fragility when the string is not exactly as anticipated, as such, we've expanded on this to specifically look for SemVer in the section of the string beginning with `aws-cli/`, ignoring whitespaces to avoid getting the version of Python, for example.

Added some bats tests to validate the behavior.